rewrite all the project

+++ roomie-server/src/main/java/com/trandnquang/roomie/dto/invoice/InvoiceDetailResponse.java
@@ -8,9 +8,9 @@ import java.math.BigDecimal;
 @Builder
 public class InvoiceDetailResponse {
     private String serviceName;
-    private Double oldIndex;
-    private Double newIndex;
-    private Double usageAmount; // Số lượng tiêu thụ
+    private BigDecimal oldIndex;
+    private BigDecimal newIndex;
+    private BigDecimal usageAmount; // Số lượng tiêu thụ
     private BigDecimal appliedPrice; // Đơn giá lúc tính
     private BigDecimal totalPrice;  // Thành tiền
     private String description;
